+++ player.js
@@ -1,6 +1,7 @@
 class Player {
 	constructor(blobManager, inputModule) {
 		this.blobManager = blobManager;
+		inputModule.attach(this); // Listen to input
 		this.inputModule = inputModule;
 
 		this.blob = null;
@@ -8,8 +9,10 @@ class Player {
 	}
 
 	spawnPlayer() {
-		this.blob = blobManager.initPlayerBlob();
-		this.blob.setInputModule(this.inputModule);
+		if (!this.blob || !this.blob.isAlive) {
+			this.blob = blobManager.initPlayerBlob();
+			this.blob.setInputModule(this.inputModule);			
+		}
 	}
 
 	update() {
@@ -18,5 +21,11 @@ class Player {
 
 	get pos() {
 		return this.blob.pos;
+	}
+
+	observerUpdate(message) {
+		if (message.message == "SpawnPlayer") {
+			this.spawnPlayer();
+		}		
 	}
 }