+++ manualinput.js
@@ -6,9 +6,18 @@ class ManualInput extends InputModule {
 
 		this.mouseIsPressedPrev = false;
 		this.keyIsPressedPrev = null;
+		this.prevKey = null;
+
+		this.spawnKey = ' ';
 	}
 
 	update() {
+		this.handelMouseInput();
+		this.handleKeyboardInput();
+		this.handleTouchInput();
+	}	
+
+	handelMouseInput() {
 		let mousePos = createVector(mouseX, mouseY);
 		if (mouseIsPressed) {
 			this.joystick.feedInput(mousePos);
@@ -18,7 +27,9 @@ class ManualInput extends InputModule {
 		}
 
 		this.mouseIsPressedPrev = mouseIsPressed;
+	}
 
+	handleTouchInput() {
 		// Might work for touch input?
 		// if (touches.length > 0) {
 		// 	if (this.previousTouchLength <= 0) {
@@ -32,30 +43,48 @@ class ManualInput extends InputModule {
 		// 	// Input stopped since last frame
 		// 	this.joystick.finishInput();
 		// }
+	}
 
-
-		// Legacy keyboard input
+	handleKeyboardInput() {
+		
 		if (keyIsPressed) {
-			if (keyCode == UP_ARROW) {
-				this.move(InputModule.VECTOR_UP);
-				// this.joystick.feedInput(p5.Vector.add(this.joystick.pos, p5.Vector.mult(InputModule.VECTOR_UP, this.joystick.radius) ));
+			if (key == this.spawnKey && this.prevKey != this.spawnKey) {
+				this.notifySpawnPlayer();
+			} else if (this.prevKey == this.spawnKey && key == this.spawnkey) {
+				// Do nothing on prolonged presses on same key.
+				// TODO: Doesn't work
+			}
+
+			// Legacy keyboard input
+			else if (keyCode == UP_ARROW) {
+				this.notifyMovement(InputModule.VECTOR_UP);
 			} else if (keyCode == DOWN_ARROW) {
-				this.move(InputModule.VECTOR_DOWN);
+				this.notifyMovement(InputModule.VECTOR_DOWN);
 			} else if (keyCode == LEFT_ARROW) {
-				this.move(InputModule.VECTOR_LEFT);
+				this.notifyMovement(InputModule.VECTOR_LEFT);
 			} else if (keyCode == RIGHT_ARROW) {
-				this.move(InputModule.VECTOR_RIGHT);
+				this.notifyMovement(InputModule.VECTOR_RIGHT);
 			} else {
 				console.log("No function for key: " + keyCode);
 			}
-		} else if (this.keyIsPressedPrev) {
-			this.joystick.finishInput();
+
+			// To prevent long unwanted long presses.
+			this.prevKey = key;
+
+		} else  {
+			this.prevKey = null;
 		}
-	}	
+	}
 
 	observerUpdate(message) {
 		if (message.message == InputEnum.MOVEMENT_VECTOR) {
-			this.move(message.vector);
+			this.notifyMovement(message.vector);
 		}
+	}
+
+	notifySpawnPlayer() {
+		this.notify({
+			message: "SpawnPlayer"
+		});
 	}
 }