+++ player.js
@@ -1,22 +1,39 @@
 class Player {
-	constructor(blobManager, inputModule) {
+	constructor(startPos, blobManager, inputModule) {
 		this.blobManager = blobManager;
+		inputModule.attach(this); // Listen to input
 		this.inputModule = inputModule;
 
 		this.blob = null;
-		this.spawnPlayer();
+		// this.lastKnownBlobPos = null;
+		this.spawnPlayer(startPos);
 	}
 
-	spawnPlayer() {
-		this.blob = blobManager.initPlayerBlob();
-		this.blob.setInputModule(this.inputModule);
+	spawnPlayer(pos) {
+		if (!this.blob || !this.blob.isAlive) {
+			this.blob = blobManager.initPlayerBlob(pos);
+			this.blob.setInputModule(this.inputModule);
+			// this.lastKnownBlobPos = this.blob.pos;			
+		}
 	}
 
 	update() {
-
+		// if (this.blob && this.blob.pos) {
+		// 	this.lastKnownBlobPos = this.blob.pos.copy();
+		// }
 	}
 
 	get pos() {
-		return this.blob.pos;
+		// if (this.blob && this.blob.pos) {
+			return this.blob.pos;
+		// } else {
+		// 	return this.lastKnownBlobPos;
+		// }
+	}
+
+	observerUpdate(message) {
+		if (message.message == "SpawnPlayer") {
+			this.spawnPlayer(this.pos);
+		}		
 	}
 }