+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Added
+- Smoothfollow camera.
+- Camera chasing faster with distance.
 
 ## [0.6.0] - 2016-12-19
 ### Added