+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Drawing gameboard edges.
 - Area outside gameboard filled.
 - Respawn player on space key down.
+- Respawning on site of demise rather than in center of gameboard.
 
 ## [0.6.0] - 2016-12-19
 ### Added