+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Added
 - Smoothfollow camera.
 - Camera chasing faster with distance.
+- Drawing gameboard edges.
+- Area outside gameboard filled.
+- Respawn player on space key down.
+- Respawning on site of demise rather than in center of gameboard.
 
 ## [0.6.0] - 2016-12-19
 ### Added